+++ dvha/models/regression.py
@@ -15,11 +15,13 @@ from pubsub import pub
 from dvha.tools.errors import ErrorDialog
 from dvha.models.plot import PlotRegression, PlotMultiVarRegression
 from dvha.models.machine_learning import RandomForestFrame, GradientBoostingFrame, DecisionTreeFrame,\
-    SupportVectorRegressionFrame
+    SupportVectorRegressionFrame, MachineLearningModelViewer
 from dvha.dialogs.export import save_data_to_file
 from dvha.dialogs.main import SelectRegressionVariablesDialog
+from dvha.options import DefaultOptions
 from dvha.paths import ICONS, MODELS_DIR
-from dvha.tools.utilities import set_msw_background_color, get_tree_ctrl_image, get_window_size
+from dvha.tools.stats import MultiVariableRegression
+from dvha.tools.utilities import set_msw_background_color, get_tree_ctrl_image, get_window_size, load_object_from_file
 
 
 class RegressionFrame:
@@ -388,7 +390,7 @@ class MultiVarResultsFrame(wx.Frame):
     """
     Class to view multi-variable regression with data passed from RegressionFrame
     """
-    def __init__(self, y_variable, x_variables, group_data, group, options):
+    def __init__(self, y_variable, x_variables, group_data, group, options, auto_update_plot=True):
         """
         :param y_variable: dependent variable
         :type y_variable: str
@@ -403,6 +405,8 @@ class MultiVarResultsFrame(wx.Frame):
 
         self.y_variable = y_variable
         self.x_variables = x_variables
+        self.group_data = group_data
+        self.group = group
         self.stats_data = group_data[group]['stats_data']
 
         set_msw_background_color(self)  # If windows, change the background color
@@ -410,17 +414,19 @@ class MultiVarResultsFrame(wx.Frame):
         self.options = options
 
         self.plot = PlotMultiVarRegression(self, options, group)
-        self.plot.update_plot(y_variable, x_variables, self.stats_data)
-        msg = {'y_variable': y_variable,
-               'x_variables': x_variables,
-               'regression': self.plot.reg,
-               'group': group}
-        pub.sendMessage('control_chart_set_model', **msg)
+        if auto_update_plot:
+            self.plot.update_plot(y_variable, x_variables, self.stats_data)
+            msg = {'y_variable': y_variable,
+                   'x_variables': x_variables,
+                   'regression': self.plot.reg,
+                   'group': group}
+            pub.sendMessage('control_chart_set_model', **msg)
 
         self.button_back_elimination = wx.Button(self, wx.ID_ANY, 'Backward Elimination')
         self.button_export = wx.Button(self, wx.ID_ANY, 'Export Plot Data')
         self.button_save_plot = wx.Button(self, wx.ID_ANY, 'Save Plot')
         self.button_save_model = wx.Button(self, wx.ID_ANY, 'Save Model')
+        self.button_load_mlr_model = wx.Button(self, wx.ID_ANY, 'Load Model')
         algorithms = ['Random Forest', 'Support Vector Machine', 'Decision Tree', 'Gradient Boosting']
         self.button = {key: wx.Button(self, wx.ID_ANY, key) for key in algorithms}
         self.radiobox_include_back_elim = wx.RadioBox(self, wx.ID_ANY, 'Include all x-variables?', choices=['Yes', 'No'])
@@ -448,6 +454,7 @@ class MultiVarResultsFrame(wx.Frame):
         self.Bind(wx.EVT_BUTTON, self.on_export, id=self.button_export.GetId())
         self.Bind(wx.EVT_BUTTON, self.on_save_plot, id=self.button_save_plot.GetId())
         self.Bind(wx.EVT_BUTTON, self.on_save_model, id=self.button_save_model.GetId())
+        self.Bind(wx.EVT_BUTTON, self.on_load_mlr_model, id=self.button_load_mlr_model.GetId())
         self.Bind(wx.EVT_SIZE, self.on_resize)
 
     def __do_layout(self):
@@ -465,6 +472,7 @@ class MultiVarResultsFrame(wx.Frame):
         sizer_algo_wrapper.Add(sizer_export_buttons, 0, wx.ALL, 5)
         text = wx.StaticText(self, wx.ID_ANY, "Compare with Machine Learning Module")
         sizer_algo_wrapper.Add(text, 0, wx.EXPAND | wx.ALL, 5)
+        sizer_algo_select.Add(self.button_load_mlr_model, 0, wx.EXPAND | wx.ALL, 5)
         for key, button in self.button.items():
             sizer_algo_select.Add(button, 0, wx.EXPAND | wx.ALL, 5)
         sizer_algo_select.Add(self.radiobox_include_back_elim, 0, wx.EXPAND | wx.ALL, 5)
@@ -514,9 +522,21 @@ class MultiVarResultsFrame(wx.Frame):
         data = {'y_variable': self.plot.y_variable,
                 'regression': self.plot.reg,
                 'x_variables': self.plot.x_variables,
-                'regression_type': 'multi-variable-linear'}
+                'regression_type': 'multi-variable-linear',
+                'version': DefaultOptions().VERSION}
         save_data_to_file(self, 'Save Model', data,
-                          wildcard="REG files (*.reg)|*.reg", data_type='pickle', initial_dir=MODELS_DIR)
+                          wildcard="MVR files (*.mvr)|*.mvr", data_type='pickle', initial_dir=MODELS_DIR)
+
+    def on_load_mlr_model(self, evt):
+        # with wx.FileDialog(self, "Load a machine learning model", "", wildcard='*.mlr',
+        #                    style=wx.FD_FILE_MUST_EXIST | wx.FD_OPEN) as dlg:
+        #     dlg.SetDirectory(MODELS_DIR)
+        #     if dlg.ShowModal() == wx.ID_OK:
+        #         model_file_path = dlg.GetPath()
+        #         saved_data = load_object_from_file(model_file_path)
+        #         print(type(saved_data['regressor']))
+        #         self.ml_frames.append(RandomForestFrame(self.final_stats_data))
+        MachineLearningModelViewer(self, self.group_data, self.group, self.options)
 
     def redraw_plot(self):
         self.plot.redraw_plot()
@@ -536,3 +556,57 @@ class MultiVarResultsFrame(wx.Frame):
                     frame.Close()
                 except RuntimeError:
                     pass
+
+
+class LoadMultiVarModelFrame(MultiVarResultsFrame):
+    def __init__(self, model_file_path, group_data, group, options):
+        self.loaded_data = load_object_from_file(model_file_path)
+        self.stats_data = group_data[group]['stats_data']
+        try:
+            if self.is_valid:
+                y_variable = self.loaded_data['y_variable']
+                x_variables = self.loaded_data['x_variables']
+                stats_data = group_data[group]['stats_data']
+
+                MultiVarResultsFrame.__init__(self, y_variable, x_variables, group_data, group, options,
+                                              auto_update_plot=False)
+                X, y = stats_data.get_X_and_y(y_variable, x_variables)
+
+                reg = MultiVariableRegression(X, y, saved_reg=self.loaded_data['regression'])
+                self.plot.update_plot(y_variable, x_variables, self.stats_data, reg=reg)
+
+                self.Show()
+            else:
+                if self.stats_data is None:
+                    msg = 'No data has been queried for Group %s.' % group
+                elif not self.is_mvr:
+                    msg = 'Selected file is not a valid multi-variable regression save file.'
+                elif not self.stats_data_has_y:
+                    msg = "The model's dependent variable is not found in your queried data:\n%s" % self.y_variable
+                elif self.missing_x_variables:
+                    msg = 'Your queried data is missing the following independent variables:\n%s' % \
+                          ', '.join(self.missing_x_variables)
+                else:
+                    msg = 'Unknown error.'
+
+                wx.MessageBox(msg, 'Model Loading Error', wx.OK | wx.OK_DEFAULT | wx.ICON_WARNING)
+        except Exception as e:
+            msg = str(e)
+            wx.MessageBox(msg, 'Model Loading Error', wx.OK | wx.OK_DEFAULT | wx.ICON_WARNING)
+
+    @property
+    def is_mvr(self):
+        return 'regression_type' in list(self.loaded_data) \
+                and self.loaded_data['regression_type'] == 'multi-variable-linear'
+
+    @property
+    def is_valid(self):
+        return self.stats_data is not None and self.is_mvr and not self.missing_x_variables and self.stats_data_has_y
+
+    @property
+    def missing_x_variables(self):
+        return [x for x in self.loaded_data['x_variables'] if x not in list(self.stats_data.data)]
+
+    @property
+    def stats_data_has_y(self):
+        return self.loaded_data['y_variable'] in list(self.stats_data.data)